Moved counting channel code to seperate file
