+++ index.js
@@ -11,18 +11,16 @@ const messageHandler = require('./handlers/messageHandler');
 const { connect } = require('mongoose');
 const docCreate = require('./events/docCreate');
 const piiUpdate = require('./events/piiUpdate');
-const piiModel = require('./models/pii');
+const gateModel = require('./models/gate');
+const counting = require('./functions/counting');
 require('dotenv').config();
 require('colors');
-global.botVersion = "1.2.15";
-// Stores the timeout used to make the bot count if nobody else counts for a set period of
-// time.
-let timeout
+global.botVersion = "1.2.16";
 
 //actions to run at bot startup
 bot.on('ready', async () => {
 	onReady.event(bot)
-	console.log("Startup script has run".red)
+	console.log("Startup script has run".red.bold)
 });
 
 //actions to run when the bot joins a server
@@ -42,32 +40,8 @@ bot.on("warn", (e) => console.warn(e));
 
 //actions to run when the bot recieves a message
 bot.on('message', async (message) => {
-	const srv = await piiModel.findOne({ GUILD_ID: message.guild.id }); //find the entry for the guild
-	// Stores the current count.
-	let count = Number(srv.GUILD_COUNTING_NUMBER);
-
-
-	// Only do this for the counting channel of course
-	if (bot.channels.cache.filter(c => c.name === 'counting').keyArray().includes(message.channel.id)) {
-		let lm;
-		// You can ignore all bot messages like this
-		if (message.member.user.bot) return
-		// If the message is the current count + 1...
-		message.channel.messages.fetch({ limit: 2 }).then(async messages => {
-			lm = messages.last();
-			if (lm.author.id == message.author.id) {
-				message.delete();
-			} else if (Number(message.content) === count + 1) {
-				// ...increase the count
-				count++
-				await piiModel.findOneAndUpdate({ GUILD_ID: message.guild.id }, { $set: { GUILD_COUNTING_NUMBER: count } }, { new: true });
-				// If the message wasn't sent by the bot...
-			} else if (message.member.id !== bot.user.id) {
-				// ...send a message because the person stuffed up the counting (and log all errors)
-				message.delete()
-			}
-		})
-	}
+	net = await gateModel.findOne({ GUILD_ID: message.guild.id });
+	counting.count(message, bot);
 	//parse commands
 	messageHandler.messageHANDLE(message, bot);
 });
@@ -82,10 +56,10 @@ bot.on('message', async (message) => {
 		useUnifiedTopology: true,
 		useFindAndModify: false
 	});
-	console.log('Connected to MongoDB'.green);
+	console.log('Connected to MongoDB'.green.bold);
 
 	//login to the discord api
 	console.log('Trying to login to the Discord API\nPlease wait for a connection'.yellow);
 	bot.login(process.env.BOT_TOKEN).catch(e => console.error(e));
-	console.log("Logged into the Discord API".green);
+	console.log("Logged into the Discord API".green.bold);
 })()
